fix: split into multiple files, fix scroll behavior on IndicatorsList
